commit five

+++ app/Http/Controllers/ListingController.php
@@ -50,6 +50,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Job created successfully');
@@ -66,6 +68,12 @@ class ListingController extends Controller
     public function update(Request $request, Listing $listing)
     {
         // dd($request->file('logo'));
+
+        // Make sure Logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -86,8 +94,20 @@ class ListingController extends Controller
     }
 
     // Delete Listing
-    public function destroy(Listing $listing) {
+    public function destroy(Listing $listing)
+    {
+        // Make sure Logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    // Manage Listings
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
